Separate precompile gas cost calculation impl from runtime impl
