+++ category/execution/CMakeLists.txt
@@ -141,6 +141,7 @@ add_library(
   "ethereum/precompiles_bls12.cpp"
   "ethereum/precompiles_bls12.hpp"
   "ethereum/precompiles_impl.cpp"
+  "ethereum/precompiles_gas_cost_impl.cpp"
   "ethereum/reserve_balance.cpp"
   "ethereum/reserve_balance.hpp"
   "ethereum/trace/call_frame.cpp"
@@ -221,7 +222,7 @@ add_library(
   "monad/min_base_fee.h"
   "monad/monad_precompiles.cpp"
   "monad/monad_precompiles.hpp"
-  "monad/monad_precompiles_impl.cpp"
+  "monad/monad_precompiles_gas_cost_impl.cpp"
   "monad/monad_transaction_gas.cpp"
   "monad/monad_transaction_gas.hpp"
   "monad/reserve_balance.cpp"
